requirements.txt
# Core dependencies
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.13
python-dotenv==1.0.0
gTTS==2.3.2
playsound==1.3.0
google-generativeai>=0.3.0

# Audio processing
webrtcvad>=2.0.10
sounddevice>=0.4.6

# Utilities
pandas>=2.0.0
matplotlib>=3.7.1
librosa>=0.10.0
numba>=0.57.0
